# docs/swagger.yaml
basePath: /v1
definitions:
  audit.Audit:
    type: object
  audit.AuditStats:
    properties:
      active_services:
        type: integer
      avg_response_time:
        type: number
      by_method:
        additionalProperties:
          type: integer
        type: object
      by_service:
        items:
          $ref: '#/definitions/audit.ServiceBreakdown'
        type: array
      by_status_class:
        additionalProperties:
          type: integer
        type: object
      errors_4xx:
        type: integer
      errors_5xx:
        type: integer
      last_event_at:
        type: string
      p95_response_time:
        type: number
      timeline:
        items:
          $ref: '#/definitions/audit.TimelinePoint'
        type: array
      total:
        type: integer
    type: object
...
  audit.ServiceBreakdown:
    properties:
      avg_response_time:
        type: number
      errors:
        type: integer
      last_event:
        type: string
      requests:
        type: integer
      service_name:
        type: string
    type: object
  audit.TimelinePoint:
    properties:
      count:
        type: integer
      hour:
        type: string
    type: object
...
host: localhost:8082
info:
  contact:
    name: BatAudit
  description: Self-hosted audit logging platform. Collect, store and query audit
    events from any service.
  license:
    name: MIT
  title: BatAudit API
  version: "1.0"
paths:
  /audit:
    get:
      description: Returns a paginated list of audit events with optional filters
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Filter by project ID
        in: query
        name: project_id
        type: string
      - description: Filter by service name
        in: query
        name: service_name
        type: string
      - description: Filter by user/client identifier
        in: query
        name: identifier
        type: string
      - description: Filter by HTTP method (GET, POST, PUT, DELETE, PATCH)
        in: query
        name: method
        type: string
      - description: Filter by HTTP status code
        in: query
        name: status_code
        type: integer
      - description: Filter by environment (prod, staging, dev)
        in: query
        name: environment
        type: string
      - description: Filter from date (ISO 8601)
        in: query
        name: start_date
        type: string
      - description: Filter to date (ISO 8601)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List audit events
      tags:
      - audit
    post:
      consumes:
      - application/json
      description: Receives an audit event from an SDK, validates and queues it for
        processing. Requires X-API-Key header.
      parameters:
      - description: API Key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Audit event
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/audit.Audit'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'BAT-001: invalid JSON / BAT-002: validation failed'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid or missing API key
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'BAT-003: queue unavailable'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ingest audit event
      tags:
      - ingest
  /audit/{id}:
    get:
      description: Returns full details of a single audit event by ID
      parameters:
      - description: Audit event UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/audit.Audit'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get audit event
      tags:
      - audit
  /audit/sessions:
    get:
      description: Returns derived user sessions using a 30-minute inactivity gap
        algorithm
      parameters:
      - description: Filter by project ID
        in: query
        name: project_id
        type: string
      - description: Filter by user/client identifier
        in: query
        name: identifier
        type: string
      - description: Filter by service name
        in: query
        name: service_name
        type: string
      - description: Filter from date (ISO 8601)
        in: query
        name: start_date
        type: string
      - description: Filter to date (ISO 8601)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List sessions
      tags:
      - audit
  /audit/stats:
    get:
      description: 'Returns aggregated metrics: totals, error rates, response times
        (avg + p95), active services, 24h timeline, breakdown by service/status/method'
      parameters:
      - description: Filter by project ID (omit for all projects)
        in: query
        name: project_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/audit.AuditStats'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Audit statistics
      tags:
      - audit
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"